requirements.txt
cryptography==36.0.2
Django==4.0.4
django-jazzmin==2.4.9
django-oauth-toolkit==2.3.0
djangorestframework==3.13.1
rsa==4.8
requests==2.27.1
gunicorn==20.1.0
whitenoise==6.2.0
OPi.GPIO==0.5.2
celery==5.3.1
redis==4.6.0